requirements.txt
pymupdf
pytesseract
Pillow
faiss-cpu
openai
langchain
streamlit
tiktoken
unidecode
python-dotenv
